Modified UI
made the bubbles its own view as the timer to fade in and out the opacity would make the bubbles jerk...  I tried a withAnimation to fade the opacity but at the end of that fade out the bubbles would stop animating.  So I took the timer completely out and made a double animation to fade out the opacity then fade it back in.  Now it works very smooth.  I also adjusted the timing a bit.